+++ src/sandbox/ai/hermit/tests/split_veg_to_seed_test.rs
@@ -34,12 +34,12 @@ fn split_veg_to_seed_test() {
     blackboard.insert(
         // I should change the GetIsInventoryGE to hold a BlackBoardValue instadt of a key
         "knife".to_owned(),
-        Variable::Chit(BlackboardValue::String("Knife".to_owned())),
+        Variable::Chit("Knife".into()),
     );
     blackboard.insert(
         // I should change the GetIsInventoryGE to hold a BlackBoardValue instadt of a key
         "stone".to_owned(),
-        Variable::Chit(BlackboardValue::String("Stone".to_owned())),
+        Variable::Chit("Stone".into()),
     );
 
     // set up  the dummy values